+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_02_24_105843) do
+ActiveRecord::Schema.define(version: 2021_02_25_171343) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -156,6 +156,7 @@ ActiveRecord::Schema.define(version: 2021_02_24_105843) do
     t.datetime "updated_at", precision: 6, null: false
     t.boolean "service_owner", default: false
     t.string "iati_reference"
+    t.string "beis_organisation_reference"
     t.index ["iati_reference"], name: "index_organisations_on_iati_reference", unique: true
   end